MongoDB_MCP_260128/requirements.txt
mcp>=1.1.0
pymongo>=4.6.0
python-dotenv>=1.0.0
